backend/pyproject.toml
[project]
name = "ltx-backend"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "psycopg2-binary",
    "httpx",
    "websockets",
    "python-dotenv",
    "python-multipart",
    "pydantic-settings",
    "boto3",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "flower>=2.0",
    "google-generativeai>=0.8",
    "Pillow>=10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
